로직 분리... 리팩토링이 너무 간절하게 필요해 ㅠㅠ

+++ ProgramCore/DAO/ContentDAO.cs
@@ -8,7 +8,7 @@ namespace ProgramCore.DAO
 {
     public class ContentDAO
     {
-        public long Uid { get; set; }
+        public int Uid { get; set; }
 
         public string HeaderTitle { get; set; }
